+++ src/driver/oracle/OracleQueryRunner.ts
@@ -17,6 +17,7 @@ import {SqlInMemory} from "../SqlInMemory";
 import {TableIndexOptions} from "../../schema-builder/options/TableIndexOptions";
 import {TableForeignKeyOptions} from "../../schema-builder/options/TableForeignKeyOptions";
 import {PromiseUtils} from "../../util/PromiseUtils";
+import {TableUnique} from "../../schema-builder/table/TableUnique";
 
 /**
  * Runs queries on a single oracle database connection.
@@ -694,6 +695,20 @@ AND cons.constraint_name = cols.constraint_name AND cons.owner = cols.owner ORDE
      * Drops a primary key.
      */
     async dropPrimaryKey(tableOrName: Table|string): Promise<void> {
+        // todo
+    }
+
+    /**
+     * Creates a new unique constraint.
+     */
+    async createUniqueConstraint(tableOrName: Table|string, uniqueConstraint: TableUnique): Promise<void> {
+        // todo
+    }
+
+    /**
+     * Drops an unique constraint.
+     */
+    async dropUniqueConstraint(tableOrName: Table|string, uniqueOrName: TableUnique|string): Promise<void> {
         // todo
     }