+++ src/index.ts
@@ -66,6 +66,7 @@ export * from "./decorator/tree/TreeLevelColumn";
 export * from "./decorator/tree/TreeParent";
 export * from "./decorator/tree/TreeChildren";
 export * from "./decorator/Index";
+export * from "./decorator/Unique";
 export * from "./decorator/Generated";
 export * from "./decorator/DiscriminatorValue";
 export * from "./decorator/EntityRepository";