+++ test/functional/query-runner/drop-index.ts
@@ -2,7 +2,7 @@ import "reflect-metadata";
 import {Connection} from "../../../src/connection/Connection";
 import {closeTestingConnections, createTestingConnections, reloadTestingDatabases} from "../../utils/test-utils";
 
-describe.only("query runner > drop index", () => {
+describe("query runner > drop index", () => {
 
     let connections: Connection[];
     before(async () => {