+++ test/functional/query-runner/entity/Post.ts
@@ -1,7 +1,9 @@
 import {Entity} from "../../../../src/decorator/entity/Entity";
 import {Column} from "../../../../src/decorator/columns/Column";
+import {Unique} from "../../../../src/decorator/Unique";
 
 @Entity()
+@Unique(["name", "text"])
 export class Post {
 
     @Column({ primary: true, unique: true })